[DEBUG] fix gcc warnings

+++ src/arstream2_rtp_sender.c
@@ -541,7 +541,7 @@ ARSTREAM2_RtpSender_t* ARSTREAM2_RtpSender_New(ARSTREAM2_RtpSender_Config_t *con
         retSender->naluCallbackUserPtr = config->naluCallbackUserPtr;
         retSender->naluFifoSize = (config->naluFifoSize > 0) ? config->naluFifoSize : ARSTREAM2_RTP_SENDER_DEFAULT_NALU_FIFO_SIZE;
         retSender->maxPacketSize = (config->maxPacketSize > 0) ? config->maxPacketSize - ARSTREAM2_RTP_TOTAL_HEADERS_SIZE : ARSTREAM2_RTP_MAX_PAYLOAD_SIZE;
-        retSender->targetPacketSize = (config->targetPacketSize > 0) ? config->targetPacketSize - ARSTREAM2_RTP_TOTAL_HEADERS_SIZE : retSender->maxPacketSize;
+        retSender->targetPacketSize = (config->targetPacketSize > 0) ? config->targetPacketSize - (int)ARSTREAM2_RTP_TOTAL_HEADERS_SIZE : retSender->maxPacketSize;
         retSender->maxBitrate = (config->maxBitrate > 0) ? config->maxBitrate : 0;
         retSender->maxLatencyMs = (config->maxLatencyMs > 0) ? config->maxLatencyMs : 0;
         retSender->maxNetworkLatencyMs = (config->maxNetworkLatencyMs > 0) ? config->maxNetworkLatencyMs : 0;
@@ -1254,11 +1254,11 @@ static int ARSTREAM2_RtpSender_SendData(ARSTREAM2_RtpSender_t *sender, uint8_t *
     struct timespec t1;
     ARSAL_Time_GetTime(&t1);
     uint64_t curTime = (uint64_t)t1.tv_sec * 1000000 + (uint64_t)t1.tv_nsec / 1000;
-    if ((maxLatencyUs > 0) && (curTime - auTimestamp > maxLatencyUs))
+    if ((maxLatencyUs > 0) && (curTime - auTimestamp > (uint64_t)maxLatencyUs))
     {
         totalLatencyDrop = 1;
     }
-    else if ((maxNetworkLatencyUs > 0) && (curTime - inputTimestamp > maxNetworkLatencyUs))
+    else if ((maxNetworkLatencyUs > 0) && (curTime - inputTimestamp > (uint64_t)maxNetworkLatencyUs))
     {
         networkLatencyDrop = 1;
     }
@@ -1307,11 +1307,11 @@ static int ARSTREAM2_RtpSender_SendData(ARSTREAM2_RtpSender_t *sender, uint8_t *
                     {
                         ARSAL_Time_GetTime(&t1);
                         curTime = (uint64_t)t1.tv_sec * 1000000 + (uint64_t)t1.tv_nsec / 1000;
-                        if ((maxLatencyUs > 0) && (curTime - auTimestamp > maxLatencyUs))
+                        if ((maxLatencyUs > 0) && (curTime - auTimestamp > (uint64_t)maxLatencyUs))
                         {
                             totalLatencyDrop = 1;
                         }
-                        else if ((maxNetworkLatencyUs > 0) && (curTime - inputTimestamp > maxNetworkLatencyUs))
+                        else if ((maxNetworkLatencyUs > 0) && (curTime - inputTimestamp > (uint64_t)maxNetworkLatencyUs))
                         {
                             networkLatencyDrop = 1;
                         }
@@ -1505,11 +1505,11 @@ void* ARSTREAM2_RtpSender_RunStreamThread (void *ARSTREAM2_RtpSender_t_Param)
 
             /* check that the NALU is not older than the max latency */
             totalLatencyDrop = networkLatencyDrop = 0;
-            if ((maxLatencyUs > 0) && (curTime - nalu.auTimestamp > maxLatencyUs))
+            if ((maxLatencyUs > 0) && (curTime - nalu.auTimestamp > (uint64_t)maxLatencyUs))
             {
                 totalLatencyDrop = 1;
             }
-            else if ((maxNetworkLatencyUs > 0) && (curTime - nalu.naluInputTimestamp > maxNetworkLatencyUs))
+            else if ((maxNetworkLatencyUs > 0) && (curTime - nalu.naluInputTimestamp > (uint64_t)maxNetworkLatencyUs))
             {
                 networkLatencyDrop = 1;
             }
@@ -1521,7 +1521,7 @@ void* ARSTREAM2_RtpSender_RunStreamThread (void *ARSTREAM2_RtpSender_t_Param)
                 fragmentCount = (nalu.naluSize + targetPacketSize / 2) / targetPacketSize;
                 if (fragmentCount < 1) fragmentCount = 1;
 
-                if ((fragmentCount > 1) || (nalu.naluSize > maxPacketSize))
+                if ((fragmentCount > 1) || (nalu.naluSize > (uint32_t)maxPacketSize))
                 {
                     /* Fragmentation (FU-A) */
 
@@ -1545,7 +1545,7 @@ void* ARSTREAM2_RtpSender_RunStreamThread (void *ARSTREAM2_RtpSender_t_Param)
 
                     for (i = 0, offset = 1; i < fragmentCount; i++)
                     {
-                        fragmentSize = (i == fragmentCount - 1) ? nalu.naluSize - offset : meanFragmentSize;
+                        fragmentSize = (i == fragmentCount - 1) ? (int)nalu.naluSize - offset : meanFragmentSize;
                         fragmentOffset = 0;
                         do
                         {
@@ -1576,8 +1576,8 @@ void* ARSTREAM2_RtpSender_RunStreamThread (void *ARSTREAM2_RtpSender_t_Param)
                 }
                 else
                 {
-                    int stapSize = nalu.naluSize + 2 + ((!stapPending) ? sizeof(ARSTREAM2_RTP_Header_t) + 1 : 0);
-                    if ((sendSize + stapSize >= maxPacketSize) || (sendSize + stapSize > targetPacketSize) || (nalu.seqNumForcedDiscontinuity))
+                    uint32_t stapSize = nalu.naluSize + 2 + ((!stapPending) ? sizeof(ARSTREAM2_RTP_Header_t) + 1 : 0);
+                    if ((sendSize + stapSize >= (uint32_t)maxPacketSize) || (sendSize + stapSize > (uint32_t)targetPacketSize) || (nalu.seqNumForcedDiscontinuity))
                     {
                         if (stapPending)
                         {
@@ -1592,7 +1592,7 @@ void* ARSTREAM2_RtpSender_RunStreamThread (void *ARSTREAM2_RtpSender_t_Param)
                         }
                     }
 
-                    if ((sendSize + stapSize >= maxPacketSize) || (sendSize + stapSize > targetPacketSize))
+                    if ((sendSize + stapSize >= (uint32_t)maxPacketSize) || (sendSize + stapSize > (uint32_t)targetPacketSize))
                     {
                         /* Single NAL unit */
                         memcpy(sendBuffer + sizeof(ARSTREAM2_RTP_Header_t), nalu.naluBuffer, nalu.naluSize);
